+++ lib/core/mock/identity-user.service.mock.ts
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-import { IdentityUserModel, IdentityJoinGroupRequestModel } from './../userinfo/models/identity-user.model';
+import { IdentityUserModel } from './../userinfo/models/identity-user.model';
 import { IdentityRoleModel } from './../userinfo/models/identity-role.model';
 import { IdentityGroupModel } from './../userinfo/models/identity-group.model';
+import { IdentityJoinGroupRequestModel } from '../services/identity-user.service';
 
 export let  mockIdentityUser1 = new IdentityUserModel(
     { id: 'mock-user-id-1', username: 'userName1', firstName: 'first-name-1', lastName: 'last-name-1', email: 'abc@xyz.com' }
@@ -68,7 +69,7 @@ export let mockEffectiveRoles = [
     new IdentityRoleModel({id: 'mock-role-id-3', name: 'MOCK-ROLE-1'})
 ];
 
-export let mockJoinGroupRequest = new IdentityJoinGroupRequestModel({userId: 'mock-hser-id', groupId: 'mock-group-id', realm: 'mock-realm-name'});
+export let mockJoinGroupRequest: IdentityJoinGroupRequestModel = {userId: 'mock-hser-id', groupId: 'mock-group-id', realm: 'mock-realm-name'};
 
 export let mockGroup1 = new IdentityGroupModel({
     id: 'mock-group-id-1', name: 'Mock Group 1', path: '/mock', subGroups: []
