+++ lib/core/services/identity-role.service.ts
@@ -19,12 +19,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { throwError as observableThrowError, Observable } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
-import {
-    AppConfigService,
-    LogService,
-    IdentityRoleModel
-} from '@alfresco/adf-core';
 import { Pagination } from '@alfresco/js-api';
+import { IdentityRoleModel } from '../userinfo/models/identity-role.model';
+import { AppConfigService } from '../app-config/app-config.service';
+import { LogService } from './log.service';
 
 export interface IdentityRoleResponseModel {
     entries: IdentityRoleModel[];
