+++ lib/core/services/identity-user.service.spec.ts
@@ -42,7 +42,7 @@ import { setupTestBed } from '../testing/setupTestBed';
 import { CoreModule } from '../core.module';
 import { AlfrescoApiService } from './alfresco-api.service';
 import { mockToken } from '../mock/jwt-helper.service.spec';
-import { IdentityUserModel, IdentityUserQueryCloudRequestModel } from '../userinfo/models/identity-user.model';
+import { IdentityUserModel } from '../userinfo/models/identity-user.model';
 import { IdentityRoleModel } from '../userinfo/models/identity-role.model';
 import { AlfrescoApiServiceMock } from '../mock/alfresco-api.service.mock';
 
@@ -292,7 +292,7 @@ describe('IdentityUserService', () => {
 
     it('should be able to query users based on query params (first & max params)', (done) => {
         spyOn(alfrescoApiService, 'getInstance').and.returnValue(queryUsersMockApi);
-        service.queryUsers(new IdentityUserQueryCloudRequestModel({first: 0, max: 5})).subscribe((res) => {
+        service.queryUsers({first: 0, max: 5}).subscribe((res) => {
             expect(res).toBeDefined();
             expect(res).not.toBeNull();
             expect(res.entries.length).toBe(5);
@@ -314,7 +314,7 @@ describe('IdentityUserService', () => {
 
         spyOn(service, 'queryUsers').and.returnValue(throwError(errorResponse));
 
-        service.queryUsers(new IdentityUserQueryCloudRequestModel({first: 0, max: 5}))
+        service.queryUsers({first: 0, max: 5})
             .subscribe(
                 () => {
                     fail('expected an error, not users');
