+++ lib/core/userinfo/models/identity-user.model.ts
@@ -15,8 +15,6 @@
  * limitations under the License.
  */
 
-import { Pagination } from '@alfresco/js-api';
-
 export class IdentityUserModel {
     id: string;
     firstName: string;
@@ -37,55 +35,6 @@ export class IdentityUserModel {
             this.createdTimestamp = obj.createdTimestamp || null;
             this.emailVerified = obj.emailVerified || null;
             this.enabled = obj.enabled || null;
-        }
-    }
-}
-
-export class IdentityUserPasswordModel {
-
-    type: string;
-    value: string;
-    temporary: boolean;
-
-    constructor(obj?: any) {
-        if (obj) {
-            this.type = obj.type;
-            this.value = obj.value;
-            this.temporary = obj.temporary;
-        }
-    }
-}
-
-export interface IdentityUserQueryResponse {
-
-    entries: IdentityUserModel[];
-    pagination: Pagination;
-}
-
-export class IdentityUserQueryCloudRequestModel {
-
-    first: number;
-    max: number;
-
-    constructor(obj?: any) {
-        if (obj) {
-            this.first = obj.first;
-            this.max = obj.max;
-        }
-    }
-}
-
-export class IdentityJoinGroupRequestModel {
-
-    realm: string;
-    userId: string;
-    groupId: string;
-
-    constructor(obj?: any) {
-        if (obj) {
-            this.realm = obj.realm;
-            this.userId = obj.userId;
-            this.groupId = obj.groupId;
         }
     }
 }
